moving all chat related UI code to the chat folder
