+++ app/src/main/java/com/example/clicker/presentation/modView/views/ModViewDragSection.kt
@@ -78,6 +78,7 @@ import com.example.clicker.presentation.stream.views.isScrolledToEnd
 import com.example.clicker.presentation.stream.views.streamManager.ModActionMessage
 import com.example.clicker.presentation.stream.views.streamManager.ModViewChat
 import com.example.clicker.presentation.stream.ClickedUIState
+import com.example.clicker.presentation.stream.views.chat.HorizontalDragDetectionBox
 import com.example.clicker.presentation.stream.views.streamManager.DetectDoubleClickSpacer
 import com.example.clicker.presentation.stream.views.streamManager.DetectDraggingOrNotAtBottomButton
 import com.example.clicker.presentation.stream.views.streamManager.DropDownMenuHeaderBox
@@ -1123,135 +1124,4 @@ import kotlin.math.roundToInt
     }
 
 
-    /**
-     * HorizontalDragDetectionBox is a [Box] that will detect the user's drag movement and will move [itemBeingDragged] accordingly. Also, depending
-     * of if the thresholds are dragged across functions such as [quarterSwipeRightAction], [quarterSwipeLeftAction] and [halfSwipeAction]
-     * once the drag stopped. Icons such as [halfSwipeIconResource], [quarterSwipeLeftIconResource] and [quarterSwipeRightIconResource] will
-     * also be shown when the user crosses those thresholds
-     *
-     * @param itemBeingDragged a composable function that will be dragged when the drags it accross the screen.
-     * @param twoSwipeOnly a boolean that is used to determine of there are functions for quarter swipes and half swipes or just quarter swipes.
-     * A true value indicates that [quarterSwipeRightAction] and [quarterSwipeLeftAction] will get triggered. A false value means that
-     * [quarterSwipeRightAction], [quarterSwipeLeftAction] and [halfSwipeAction] will get triggered
-     * @param quarterSwipeRightAction is a function that will be called if a user swipes and passes the threshold of 0.25 of [itemBeingDragged] width
-     * @param quarterSwipeLeftAction is a function that will be called if a user swipes and passes the threshold of -1*(0.25) of [itemBeingDragged] width
-     * @param halfSwipeAction a optional function that will be called if [twoSwipeOnly] is set to false and the user's drag passes
-     * the threshold of +/- 0.5 of [itemBeingDragged] width
-     * @param halfSwipeIconResource is a [Painter] that will be shown to the user if the half swipe threshold is crossed and [twoSwipeOnly] is false
-     * @param quarterSwipeLeftIconResource is a [Painter] that will be shown to the user if the -1 *(quarter) swipe threshold is crossed
-     * @param quarterSwipeRightIconResource is a [Painter] that will be shown to the user if the quarter swipe threshold is crossed
-     * @param hideIconColor: a [Color] that the icons will be set to hide them from the user
-     * @param showIconColor: a [Color] that the icons will be set to reveal them to the user
-     * */
-    @Composable
-    fun HorizontalDragDetectionBox(
-        itemBeingDragged:@Composable (dragOffset:Float) -> Unit,
-        twoSwipeOnly:Boolean,
-        quarterSwipeRightAction:()->Unit,
-        quarterSwipeLeftAction:()->Unit,
-        halfSwipeAction:()->Unit={},
-        halfSwipeIconResource: Painter = painterResource(id = R.drawable.delete_outline_24),
-        quarterSwipeLeftIconResource: Painter = painterResource(id = R.drawable.time_out_24),
-        quarterSwipeRightIconResource: Painter = painterResource(id = R.drawable.ban_24),
-        hideIconColor: Color = MaterialTheme.colorScheme.primary,
-        showIconColor: Color = MaterialTheme.colorScheme.onPrimary,
-        swipeEnabled:Boolean
-    ){
-        var iconPainterResource: Painter = painterResource(id = R.drawable.ban_24)
-        var dragging by remember{ mutableStateOf(true) }
-        val state = rememberDraggableActions()
-        val offset = if(swipeEnabled) state.offset.value else 0f
-        var iconColor = hideIconColor
 
-        if(dragging && !twoSwipeOnly){
-            if (state.offset.value >= (state.halfWidth)) {
-                iconPainterResource =halfSwipeIconResource
-                iconColor = showIconColor
-            }
-            else if (state.offset.value <= -(state.halfWidth)){
-                iconPainterResource =halfSwipeIconResource
-                iconColor = showIconColor
-            }
-            else if (state.offset.value <= -(state.quarterWidth)){
-                iconPainterResource =quarterSwipeLeftIconResource
-                iconColor = showIconColor
-            }
-            else if (state.offset.value >= (state.quarterWidth)){
-                iconPainterResource = quarterSwipeRightIconResource
-                iconColor = showIconColor
-            }
-        }
-        else if(dragging && twoSwipeOnly){
-            if (state.offset.value <= -(state.quarterWidth)){
-                iconPainterResource =quarterSwipeLeftIconResource
-                iconColor = showIconColor
-            }
-            else if (state.offset.value >= (state.quarterWidth)){
-                iconPainterResource = quarterSwipeRightIconResource
-                iconColor = showIconColor
-            }
-        }
-
-
-        Box(
-            modifier = Modifier
-                .fillMaxWidth()
-                .draggable(
-                    orientation = Orientation.Horizontal,
-                    onDragStopped = {
-                        if (twoSwipeOnly && swipeEnabled) {
-                            state.checkQuarterSwipeThresholds(
-                                leftSwipeAction = {
-                                    quarterSwipeLeftAction()
-                                },
-                                rightSwipeAction = {
-                                    quarterSwipeRightAction()
-                                }
-                            )
-                        } else if (swipeEnabled) {
-                            state.checkDragThresholdCrossed(
-                                deleteMessageSwipe = {
-                                    halfSwipeAction()
-                                },
-                                timeoutUserSwipe = {
-                                    quarterSwipeLeftAction()
-                                },
-                                banUserSwipe = {
-                                    quarterSwipeRightAction()
-                                }
-                            )
-                        }
-
-                        dragging = false
-                        state.resetOffset()
-                    },
-                    onDragStarted = {
-                        dragging = true
-                    },
-
-
-                    enabled = true,
-                    state = state.draggableState
-                )
-                .onGloballyPositioned { layoutCoordinates ->
-                    state.setWidth(layoutCoordinates.size.width)
-                }
-        ){
-
-            Icon(painter = iconPainterResource, contentDescription = "",tint = iconColor, modifier = Modifier
-                .align(Alignment.CenterEnd)
-                .padding(end = 10.dp)
-            )
-            Icon(painter = iconPainterResource, contentDescription = "",tint = iconColor,
-                modifier = Modifier
-                    .align(Alignment.CenterStart)
-                    .padding(start = 10.dp)
-            )
-
-
-            itemBeingDragged(offset)
-
-        }
-
-
-    }
