+++ app/src/main/java/com/example/clicker/presentation/stream/views/chat/ChatMessages.kt
@@ -1,6 +1,5 @@
-package com.example.clicker.presentation.sharedViews
+package com.example.clicker.presentation.stream.views.chat
 
-import android.util.Log
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
@@ -30,12 +29,10 @@ import androidx.compose.ui.graphics.painter.Painter
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.res.painterResource
-import androidx.compose.ui.text.SpanStyle
-import androidx.compose.ui.text.buildAnnotatedString
-import androidx.compose.ui.text.withStyle
 import androidx.compose.ui.unit.TextUnit
 import androidx.compose.ui.unit.dp
 import com.example.clicker.R
+import com.example.clicker.presentation.sharedViews.fadingEdge
 
 
 /**
